Skyrim support

+++ Oscape_CBash-bsa.C
@@ -1,3 +1,33 @@
+/* Version: MPL 1.1/LGPL 3.0
+ *
+ * "The contents of this file are subject to the Mozilla Public License
+ * Version 1.1 (the "License"); you may not use this file except in
+ * compliance with the License. You may obtain a copy of the License at
+ * http://www.mozilla.org/MPL/
+ *
+ * Software distributed under the License is distributed on an "AS IS"
+ * basis, WITHOUT WARRANTY OF ANY KIND, either express or implied. See the
+ * License for the specific language governing rights and limitations
+ * under the License.
+ *
+ * The Original Code is BSAopt.
+ *
+ * The Initial Developer of the Original Code is
+ * Ethatron <niels@paradice-insight.us>. Portions created by The Initial
+ * Developer are Copyright (C) 2011 The Initial Developer.
+ * All Rights Reserved.
+ *
+ * Alternatively, the contents of this file may be used under the terms
+ * of the GNU Library General Public License Version 3 license (the
+ * "LGPL License"), in which case the provisions of LGPL License are
+ * applicable instead of those above. If you wish to allow use of your
+ * version of this file only under the terms of the LGPL License and not
+ * to allow others to use your version of this file under the MPL,
+ * indicate your decision by deleting the provisions above and replace
+ * them with the notice and other provisions required by the LGPL License.
+ * If you do not delete the provisions above, a recipient may use your
+ * version of this file under either the MPL or the LGPL License."
+ */
 
 #define REMOVE_DOUBLES
 #define	DEPEXT_ZLIB
@@ -397,7 +427,7 @@ public:
 
 	  if (fread(&header, 1, sizeof(header), ibsa) != sizeof(header))
 	    return shutdown("Can't read from the BSA!");
-	  if (fseek(ibsa, 0, SEEK_END))
+	  if (fseek(ibsa, 0L, SEEK_END))
 	    return shutdown("Can't read from the BSA!");
 
 	  sizefile = ftell(ibsa);
@@ -694,7 +724,7 @@ public:
 		_header.FolderRecordOffset +
 		_header.FolderCount * sizeof(OBBSAFolderInfo) +
 		_header.FileNameLength +
-		(dfmix - &dfblob[0]);
+		(unsigned int)(dfmix - &dfblob[0]);
 
 	      /* assign known data */
 //	      memcpy(dinfos, &dir, sizeof(dir));
@@ -1729,7 +1759,7 @@ public:
       return file->ina;
   }
 
-  int seek(const bsfile *file, size_t offset, int origin) {
+  int seek(const bsfile *file, long offset, int origin) {
     int ola = 0, pos;
 
     /**/ if (file->oup) {
@@ -2138,7 +2168,7 @@ bool __stdcall feof_arc(void *file) {
   return r->arc->eof(r->file);
 }
 
-int __stdcall fseek_arc(void *file, size_t offset, int origin) {
+int __stdcall fseek_arc(void *file, long offset, int origin) {
   it_file_arc *r = (it_file_arc *)file;
   return r->arc->seek(r->file, offset, origin);
 }
